refactor(map): modularize map components into separate files
Separate map screen concerns into dedicated modules for better maintainability:
- Extract UI dimension constants to MapConstants.kt
- Move event marker rendering logic to EventMarkers.kt
- Simplify Map.kt to focus on screen composition

This improves code organization and makes the codebase more navigable.
